rewrite fetchQueuedFlows and fetchRecentlyFinishedFlows in JdbcExecutorLoader
rewrite fetchQueuedFlows and fetchRecentlyFinishedFlows methods to
leverage new DB API. Also added related test methods.
